refactor(request): extract pull request author resolution
